fix: dead letter error serialization
Axios errors have a lot of nested properties and are usally huge objects
which causes huge messages to be processed by RabbitMQ (several MB) as
well as crashloops of service-dead-letter which is unable to persist the
dead-letter in MongoDB because the JSON document is too big.

including the error in context is useful only for logging purposes:
since the dead-letter already includes the transaction id, including
non-enumerable properties of errors is redundant.

By going through `JSON.stringify`, we have better chances that the
library will have thrown a value that is not too big, by implementing
the `.toJSON` method like axios errors do.

TMS-345